Bugfixes
Fixes two bugs: 1) As the config is stringified & parsed the RegExp instance is lost. Creating a new RegExp instance from the config string fixes this. 2) The prefix "/base" is stripped from dep paths which leaves the paths starting with a "/". Requirejs resolves these paths without baseUrl so the modules are not found. Adding a "/" to the baseUrl (If not already there) fixes this.